+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.7.2] - 2023-03-20
+### Changed
+- Better support for historical PacBio RSII reads.
+- Cleaner TE output in the final annotation GFF3 file.
+
 ## [1.7.1] - 2023-01-28
 ### Fixed
 - A bug leading to the loss of detailed TE subtype info in the final annotation GFF3 file.